refactor: use existing instance utilities for links
