+++ apps/builder/app/builder/shared/instance-context-menu.tsx
@@ -33,9 +33,9 @@ import {
   type InstancePath,
 } from "@webstudio-is/project-build/runtime";
 import { canDeleteInstanceInContentMode } from "@webstudio-is/project-build/runtime";
-import { getInstanceLink } from "~/shared/pages/instance-link";
+import { getInstanceLink } from "~/shared/instance-utils/link";
 import { $instances, $pages, $project } from "~/shared/sync/data-stores";
-import { getDeepLinkedInstanceSelection } from "~/shared/pages/instance-link-utils";
+import { getDeepLinkedInstanceSelection } from "~/shared/instance-utils/selection";
 import {
   isComponentDetachable,
   ROOT_INSTANCE_ID,
